Send request to server to get game status
The application can post to the server at http://localhost:5000/game, which
returns a game status. The UI will update based on the game status--the label
will display a winner or a tie, and the buttons will be disabled if the game
is over.

The application currently uses a network manager class to make network calls,
and prints the response to the console in order to debug.

Commits included:

Added BoardArray, and tags to UIButtons

Add NetworkManager to make network calls to server

Post to server from GameViewController

GameViewController integrates with NetworkManager to post the game
state to the server, and it currently expects the server to return
the "true" or "false" depending on whether the game has a winner.

Display label for tie and re-factor