Computer turn and HTTPClient refactor
This Re-factors HTTP Post request to take custom completion handler. The HTTPClient
still requires a while loop to wait for request to finish. Minor changes to enable a
player to play against the computer. View now includes an activity indicator to indicate
when a move is in-progress.

TO DO: If necessary, the computer should sleep for 1 second so the move doesn't appear
at the exact same time as the human move. This will be made easier if we can
figure out how to optimize the POST request so it doesn't lock the program waiting
for the response to come back.